Module btc_relay::pallet::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

BestBlock

Store the current blockchain tip

BestBlockHeight

Store the height of the best block

BlockHeaders

Store Bitcoin block headers

ChainCounter

Increment-only counter used to track new BlockChain entries

Chains

Priority queue of BlockChain elements, ordered by the maximum height (descending). The first index into this mapping (0) is considered to be the longest chain. The value of the entry is the index into ChainsIndex to retrieve the BlockChain.

ChainsHashes

Stores a mapping from (chain_index, block_height) to block hash

ChainsIndex

Auxiliary mapping of chains ids to BlockChain entries. The first index into this mapping (0) is considered to be the Bitcoin main chain.

DisableDifficultyCheck

Whether the module should perform difficulty checks.

DisableInclusionCheck

Whether the module should perform inclusion checks.

StableBitcoinConfirmations

Global security parameter k for stable Bitcoin transactions

StableParachainConfirmations

Global security parameter k for stable Parachain transactions

StartBlockHeight

BTC height when the relay was initialized