Crate btc_relay

BTC-Relay Pallet

Based on the specification.

This pallet implements a Bitcoin light client to store and verify block headers in accordance with SPV assumptions - i.e. longest chain.

Unless otherwise stated, the primary source of truth for code contained herein is the Bitcoin Core repository, though implementation details may vary.

Overview

The BTC-Relay pallet provides functions for:

Initializing and updating the relay.

Transaction inclusion verification.

Transaction validation.

Terminology

Bitcoin Confirmations: The minimum number of Bitcoin confirmations a Bitcoin block header must have to be seen as included in the main chain.

Parachain Confirmations: The minimum number of Parachain confirmations a Bitcoin block header must have to be usable in transaction inclusion verification.

Re-exports

pub use bitcoin;

pub use types::OpReturnPaymentData;

pub use types::RichBlockHeader;

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

types

Structs

BtcPublicKey

Compressed ECDSA (secp256k1 curve) Public Key

Enums

BtcAddress

A Bitcoin address is a serialized identifier that represents the destination for a payment. Address prefixes are used to indicate the network as well as the format. Since the Parachain follows SPV assumptions we do not need to know which network a payment is included in.

Constants

ACCEPTED_MAX_TRANSACTION_OUTPUTS

ACCEPTED_MIN_TRANSACTION_OUTPUTS

ACCEPTED_MIN_TRANSACTION_OUTPUTS_WITH_OP_RETURN

ACCEPTED_NO_TRANSACTION_OUTPUTS

Number of outputs expected in the accepted transaction format See: https://interlay.gitlab.io/polkabtc-spec/btcrelay-spec/intro/accepted-format.html

DIFFICULTY_ADJUSTMENT_INTERVAL

Difficulty Adjustment Interval

MAIN_CHAIN_ID

Main chain id

TARGET_SPACING

Target Spacing: 10 minutes (600 seconds)

TARGET_TIMESPAN

Target Timespan: 2 weeks (1209600 seconds)

TARGET_TIMESPAN_DIVISOR

UNROUNDED_MAX_TARGET

Unrounded Maximum Target 0x00000000FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF

Traits

WeightInfo

Weight functions needed for btc_relay.